frontend/requirements.txt
Flask==2.0.2
azure-servicebus